+++ modules/generate_capability.go
@@ -1,0 +1,25 @@
+package modules
+
+import "strings"
+
+var (
+	capabilityVarsTfFilename = `variables.tf`
+	capabilityVarsTf         = `variable "app_metadata" {
+  description = <<EOF
+Nullstone automatically injects metadata from the app module into this module through this variable.
+This variable is a reserved variable for capabilities.
+EOF
+
+  type    = map(string)
+  default = {}
+}
+`
+)
+
+func generateCapability(manifest *Manifest) error {
+	if !strings.HasPrefix(manifest.Category, "capability/") {
+		// We don't generate capability tf if not a capability module
+		return nil
+	}
+	return generateFile(capabilityVarsTfFilename, capabilityVarsTf)
+}